chatbot/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Your project description"
authors = [
    {name = "Your Name"}
]

packages = [
    { include = "app" },
    { include = "pyproject.toml" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
crewai = "^0.108.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
pyyaml = "^6.0.2"
crewai-tools = ">=0.32.1"
asyncpg = "^0.30.0"
supabase = "^2.11.0"
langchain = "^0.3.15"
fake-useragent = "^2.0.3"
google-search-results = "2.4.2"
nltk = "^3.8.1"
extruct = "0.13.0"
w3lib = "^2.1.2"
lxml = "4.9.3"
selenium = "^4.17.2"
webdriver-manager = "^4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"